+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support a function for `debug` option, i.e. `query('...', { debug: (name, value) => /* ... */ })`
 - Disallowed whitespace between `$` and identifier, now it throws a parse error
+- Allowed parent's scope variables overlapping, i.e. `$a;.($a; ...)` doesn't throw an error now
 - Reworked build:
     - Added baking `src/parser.js` before publishing, i.e. replace runtime parser compilation with a compiled version
     - Moved `jison` to dev dependencies, package has no dependencies now (dev only)